added date checks to  scheduledPosts.js

+++ deploy/schedulePosts.js
@@ -4,35 +4,43 @@
 
 const fs = require('fs');
 const path = require('path');
+const matter = require('gray-matter');
 
-// Define the paths 
+// relative paths
 const scriptDir = path.resolve(__dirname);
-const publishDir = path.resolve(scriptDir, '..'); 
-const scheduledDir = path.resolve(scriptDir, '../scheduled'); // Scheduled dir
+const publishDir = path.resolve(scriptDir, '..'); // root
+const scheduledDir = path.resolve(scriptDir, '../scheduled'); // scheduled
 
 console.log(`Script Directory: ${scriptDir}`);
 console.log(`Publish Directory: ${publishDir}`);
 console.log(`Scheduled Directory: ${scheduledDir}`);
 
 try {
-
+  
   const files = fs.readdirSync(scheduledDir);
   console.log(`Scheduled files: ${files.join(', ')}`);
 
   files.forEach(file => {
     const filePath = path.join(scheduledDir, file);
     const content = fs.readFileSync(filePath, 'utf8');
-
-    const publishDate = new Date(content.publishDate);
+    
+    const parsedContent = matter(content);
+    const publishDate = new Date(parsedContent.data.publishDate);
+    
     console.log(`Publish Date: ${publishDate}, Current Date: ${new Date()}`);
 
-    if (new Date() >= publishDate) {
-      const publishPath = path.join(publishDir, file);
-      fs.writeFileSync(publishPath, content);
-      fs.unlinkSync(filePath);
-      console.log(`Published ${file}`);
+    if (!isNaN(publishDate)) {
+      // check dates
+      if (new Date() >= publishDate) {
+        const publishPath = path.join(publishDir, file);
+        fs.writeFileSync(publishPath, parsedContent.content);
+        fs.unlinkSync(filePath);
+        console.log(`Published ${file}`);
+      } else {
+        console.log(`Not yet time to publish ${file}`);
+      }
     } else {
-      console.log(`Not yet time to publish ${file}`);
+      console.log(`Invalid publish date for ${file}`);
     }
   });
 } catch (error) {
